refactor: ウィンドウ生成の共通化と SessionList の D&D ロジック抽出（監査・低）
- windows/shared.ts を追加し、popover/settings/setup で重複していた
  webPreferences と dev/prod ロード分岐を共通化
- SessionList の D&D 並び替え（ページ跨ぎ・drop/dragend の一度きり確定を含む）を
  useDragReorder フックに抽出し、SessionList の責務を縮小

挙動は不変（既存テスト 519 件パス、ページ跨ぎ並び替えテスト含む）。
App.tsx の prop drilling は純粋な drilling ではないと判断し今回は見送り。
