+++ src/main/windows/settings.ts
@@ -1,6 +1,6 @@
 import { BrowserWindow } from 'electron'
-import { join } from 'path'
 import { broadcastThemeOnLoad } from './themeBroadcast'
+import { sharedWebPreferences, loadRenderer } from './shared'
 import type { SettingsStore } from '../settingsStore'
 
 let settingsWindow: BrowserWindow | null = null
@@ -16,20 +16,11 @@ export function createSettingsWindow(settingsStore: SettingsStore): void {
     height: 500,
     resizable: false,
     title: 'Juice 設定',
-    webPreferences: {
-      preload: join(__dirname, '../preload/index.js'),
-      contextIsolation: true,
-      nodeIntegration: false,
-    },
+    webPreferences: sharedWebPreferences,
   })
 
   broadcastThemeOnLoad(settingsWindow, settingsStore)
-
-  if (process.env['NODE_ENV'] === 'development') {
-    settingsWindow.loadURL('http://localhost:5174/#settings')
-  } else {
-    settingsWindow.loadFile(join(__dirname, '../renderer/index.html'), { hash: 'settings' })
-  }
+  loadRenderer(settingsWindow, 'settings')
 
   settingsWindow.on('closed', () => {
     settingsWindow = null
